backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0

langchain==0.1.0
langchain-community==0.0.10

qdrant-client>=1.7.1

openai==1.6.1
httpx==0.25.2

python-dotenv==1.0.0
python-multipart==0.0.6

pydantic==2.5.0
pydantic-settings==2.1.0

pymupdf==1.24.9

redis==5.0.1
apscheduler==3.10.4

beautifulsoup4==4.12.2
pypdf2==3.0.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

websockets==12.0

httpx
requests
beautifulsoup4
lxml
python-dotenv

pillow
pytesseract
fitz